feat(xp): getting new xp after changes

+++ app/src/main/java/com/shapeup/api/services/users/UsersApi.kt
@@ -141,7 +141,7 @@ class UsersApi(
         return when (response?.status) {
             HttpStatusCode.OK -> {
                 return GetUserXpStatement(
-                    data = response.body<Long>(),
+                    data = response.body<Int>(),
                     status = response.status
                 )
             }
